* develop: (471 commits)
  [HOT FIX]  Fixing a bunch of unit tests that I broke.  My bad.
  [HOT FIX] Patch to SimulationBaseClass so that we can log Eigen matrices.
  [BSK-408] These variables changed type and name so the test was failing
  [BSK-408] Moved the location of linkInStates for spaceCraftPlus
  [BSK-408] Setting up the test dynamics unit test and hingedRigidBodyUnit test to make it so they are set up with the same initial conditions
  [BSK-408] Did the same thing for wheels and panels? Horrible
  [BSK-408] Whoops, I accidentally put this in the destructor.
  [BSK-408] Included the ability to change the name of each stateEffector. It is necessary if you have multiple instantiations of a stateEffector
  [BSK-408] Forgot to add rPrimeCB_B to the stateEffector mass prop rates
  [BSK-390] MRP Switching is now in Basilisk
  [BSK-408] Made quite a few name changes and added hubEffector to compute mass props instead of having that inside of hubEffector computeDerivatives
  [BSK-408] Added capability of stateEffector producing an rPrimeCB_B. Need this for cPrime_B
  [BSK-408] Needed to add up mass prop and mass prop rates for stateEffectors into the spaceCraft mass props and mass prop rates
  [BSK-408] Forgot that we needed cPrime_B so needed to add appropriate variable for stateEffector
  [BSK-408] Was trying to access index (2,2) of a Vector3d, that won't work.
  [HOT FIX] Removing unnecessary variables after converting interface.
  [HOT FIX] Converting DV guidance over to use the standard attitude reference output structure.
  [BSK 408] Creating single addStateEffector method that allows all stateEffectors to be linked to dynamics.
  [HOTFIX] Variable name changes for effProps
  [BSK-388] Everything compiles. Thats's good!
  ...


Former-commit-id: df8f39e6526c9b72ff210cca6924c629edbd090e
Former-commit-id: f4e2bb6e5a10ed1f97e269c304274c0768095df5

+++ SimCode/utilities/gauss_markov.cpp
@@ -25,7 +25,7 @@ GaussMarkov::GaussMarkov()
 {
     std::normal_distribution<double>::param_type
     UpdatePair(0.0, 1.0/3.0);
-    rGen.seed(RNGSeed);
+    rGen.seed((unsigned int)RNGSeed);
     rNum.param(UpdatePair);
 }
